API_books/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.3.1
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
urllib3==2.0.2
